+++ app/src/main/java/org/android/go/sopt/data/Api/HomeApiService.kt
@@ -1,7 +1,7 @@
 package org.android.go.sopt.data.Api
 
 
-import org.android.go.sopt.data.model.UserList
+import org.android.go.sopt.data.model.home.UserList
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
