+++ app/src/main/java/org/android/go/sopt/data/model/home/ResponseUserInfo.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.data.model
+package org.android.go.sopt.data.model.home
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
