+++ app/src/main/java/org/android/go/sopt/data/model/sign/RequestSignUpDto.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.data.model
+package org.android.go.sopt.data.model.sign
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
