+++ app/src/main/java/org/android/go/sopt/data/repoImpl/SignRepoImpl.kt
@@ -2,9 +2,9 @@ package org.android.go.sopt.data.repoImpl
 
 import org.android.go.sopt.domain.SignRepository
 import org.android.go.sopt.data.Api.SignApiService
-import org.android.go.sopt.data.model.RequestSignInDto
-import org.android.go.sopt.data.model.RequestSignUpDto
-import org.android.go.sopt.data.model.ResponseSignUpDto
+import org.android.go.sopt.data.model.sign.RequestSignInDto
+import org.android.go.sopt.data.model.sign.RequestSignUpDto
+import org.android.go.sopt.data.model.sign.ResponseSignUpDto
 import retrofit2.Response
 import javax.inject.Inject
