+++ app/src/main/java/org/android/go/sopt/domain/HomeRepository.kt
@@ -2,7 +2,7 @@ package org.android.go.sopt.domain
 
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
-import org.android.go.sopt.data.model.ResponseUserInfo
+import org.android.go.sopt.data.model.home.ResponseUserInfo
 
 interface HomeRepository {
     fun getUserList(): Flow<PagingData<ResponseUserInfo>>
