+++ app/src/main/java/org/android/go/sopt/ui/home/HomeViewModel.kt
@@ -3,7 +3,7 @@ package org.android.go.sopt.ui.home
 import androidx.lifecycle.ViewModel
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
-import org.android.go.sopt.data.model.ResponseUserInfo
+import org.android.go.sopt.data.model.home.ResponseUserInfo
 import org.android.go.sopt.domain.HomeRepository
 import javax.inject.Inject
