+++ app/src/main/java/org/android/go/sopt/ui/join/JoinViewModel.kt
@@ -6,9 +6,11 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
+import org.android.go.sopt.App
+import org.android.go.sopt.data.model.UserInfo
 import org.android.go.sopt.domain.SignRepository
-import org.android.go.sopt.data.model.RequestSignInDto
-import org.android.go.sopt.data.model.RequestSignUpDto
+import org.android.go.sopt.data.model.sign.RequestSignInDto
+import org.android.go.sopt.data.model.sign.RequestSignUpDto
 import org.android.go.sopt.util.Constants.ID_COUNT_MAX
 import org.android.go.sopt.util.Constants.ID_COUNT_MIN
 import org.android.go.sopt.util.Constants.ID_REGEX_MSG
